Saving the Entity Framework method, Error on connection, EntityFrameworkCore.Database.Connection[20004], no solutions found, considering to try Drapper library

+++ TrafficPoliceBlazor/TrafficPoliceBlazor/Data/cars.cs
@@ -5,10 +5,8 @@
         public string NumberPlate { get; set; }
         public string Brand { get; set; }
         public string Model { get; set; }
-
         public string Colour { get; set; }
-
-        public int Owner { get; set; }
+        public long Owner { get; set; }
 
     }
 }
